renamed processors module name
