+++ clayrs/content_analyzer/information_processor/ekphrasis_processor.py
