+++ test/content_analyzer/embeddings/embedding_learner/test_embedding_learner.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import os
 
-from clayrs.content_analyzer.information_processor.nltk import NLTK
+from clayrs.content_analyzer.information_processor.nltk_processor import NLTK
 from clayrs.content_analyzer.raw_information_source import JSONFile
 from test.content_analyzer.embeddings.test_embedding_source import TestEmbeddingSource
 from clayrs.content_analyzer.embeddings.embedding_learner.doc2vec import GensimDoc2Vec
