+++ test/content_analyzer/information_processor/test_spacy.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from clayrs.content_analyzer.information_processor.spacy import Spacy
+from clayrs.content_analyzer.information_processor.spacy_processor import Spacy
 
 
 class TestSpacy(TestCase):
